pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-agent"
version = "0.1.0"
description = "A Databricks AgentBricks agent"
requires-python = ">=3.10"
dependencies = [
    "databricks-agentbricks>=0.1.0",
    "databricks-sdk>=0.10.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "langchain>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.20.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
